Binary player not quite working for some reason

+++ cpp_graph_game/CrazyAra/main/playmode.cpp
@@ -8,18 +8,20 @@
 #include <unistd.h>
 
 void playmode(MCTSAgent * mctsAgent, RawNetAgent * rawAgent, SearchLimits * searchLimits, EvalInfo * evalInfo){
-	string action, command, subcommand;
+	string action, command, subcommand, sgf, argument, sgf_path;
 	torch::Tensor policy;
 	int move, response, hex_size;
 	istringstream iss;
 	stringstream ss;
 	bool show = false;
+	bool autoeval = true;
+	bool nexteval = true;
 	bool use_mcts = false;
 	std::vector<torch::Tensor> inputs, outputs;
 	hex_size = Options["Hex_Size"];
 	unique_ptr<Node_switching_game> game = std::make_unique<Node_switching_game>(hex_size,Options["Swap_Allowed"]);
 	while (true){
-		if (game->who_won()==NOPLAYER){
+		if (game->who_won()==NOPLAYER && nexteval){
 			if (use_mcts){
 				mctsAgent->set_search_settings(game.get(),searchLimits,evalInfo);
 				mctsAgent->evaluate_board_state();
@@ -49,6 +51,7 @@ void playmode(MCTSAgent * mctsAgent, RawNetAgent * rawAgent, SearchLimits * sear
 			cout << endl;
 		}
 #ifndef NO_PLAY
+		cout << "onturn: " << (game->onturn==RED?"red":"blue") << endl;
 		cout << "board_moves_red: ";
 		for (int i=0;i<game->board_moves_red.size();++i){
 			cout << game->board_moves_red[i] << " ";
@@ -60,26 +63,50 @@ void playmode(MCTSAgent * mctsAgent, RawNetAgent * rawAgent, SearchLimits * sear
 		}
 		cout << endl;
 #endif
-		vector<string> nodetext(game->graph.num_vertices);
-		for (int i=2;i<game->graph.num_vertices;++i){
-			ss.str(string());
-			ss << game->graph.lprops[BOARD_LOCATION][i] << "(" << setprecision(3) << policy[i-2].item<double>() << ")";
-			nodetext[i] = ss.str();
-		}
-		game->graphviz_me(nodetext,"my_graph.dot",game->graph);
-		system("neato -Tpdf my_graph.dot -o my_graph.pdf");
-		if (show){
-			system("pkill -f 'mupdf my_graph.pdf'");
-			system("mupdf my_graph.pdf &");
-			usleep(100000U);
-			system("bspc node -f west");
+		if (nexteval){
+			vector<string> nodetext(game->graph.num_vertices);
+			for (int i=2;i<game->graph.num_vertices;++i){
+				ss.str(string());
+				ss << game->graph.lprops[BOARD_LOCATION][i] << "(" << setprecision(3) << policy[i-2].item<double>() << ")";
+				nodetext[i] = ss.str();
+			}
+			game->graphviz_me(nodetext,"my_graph.dot",game->graph);
+			system("neato -Tpdf my_graph.dot -o my_graph.pdf");
+			if (show){
+				system("pkill -f 'mupdf my_graph.pdf'");
+				system("mupdf my_graph.pdf &");
+				usleep(100000U);
+				system("bspc node -f west");
+			}
 		}
 		move = -1;
 		cout << "readyok" << endl;
 		std::getline(cin,command);
 		iss = istringstream(command);
 		getline(iss,action,' ');
+		nexteval = autoeval;
 		if (action == "show") show=!show;
+		else if (action == "set_autoeval"){
+			getline(iss,argument,' ');
+			if (argument == "true"){
+				autoeval = true;
+			}
+			else{
+				autoeval = false;
+			}
+		}
+		else if (action == "eval"){
+			nexteval = true;
+		}
+		else if (action == "sgf"){
+			cout << "Loading sgf" << endl;
+			getline(iss,sgf_path,' ');
+			ifstream sgffile(sgf_path);
+			getline(sgffile,sgf);
+			game->load_sgf(sgf);
+			sgffile.close();
+			hex_size = game->board_size;
+		}
 		else if (action == "mcts") use_mcts=true;
 		else if (action == "raw") use_mcts=false;
 		else if (action == "switch") game->switch_onturn();
